chore(mobile): Move proposal related mutations to PostBodyProposal component
